requirements.txt
Flask
pillow
numpy
onnxruntime